支持ie8，增加防止csrf漏洞造成的安全问题

+++ src/main/java/com/fruit/pitaya/service/StockService.java
@@ -26,7 +26,7 @@ public class StockService {
     }
 
     @Transactional
-    public void update(String sku, Long count) {
+    public void update(String sku, int count) {
         jdbcTemplate.update("UPDATE sk_stock SET quantity=quantity-? WHERE sku=?", new Object[]{count, sku});
     }
 }
